refactor: GameCommand 패키지 이동
