+++ src/main/java/chess/domain/chessgame/gamecommand/GameCommand.java
@@ -1,4 +1,4 @@
-package chess.domain.chessgame;
+package chess.domain.chessgame.gamecommand;
 
 import java.util.Arrays;
