+++ src/main/java/chess/domain/chessgame/gamecommand/GameCommandExecutor.java
@@ -1,4 +1,4 @@
-package chess.domain.chessgame;
+package chess.domain.chessgame.gamecommand;
 
 import chess.domain.chessboard.ChessBoard;
 import java.util.List;
