+++ src/test/java/chess/domain/chessgame/GameCommandTest.java
@@ -1,8 +1,9 @@
 package chess.domain.chessgame;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertAll;
 
+import chess.domain.chessgame.gamecommand.GameCommand;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
